Add and style user info in popup
